+++ src/core/FtpFileSystem.h
@@ -188,7 +188,7 @@ protected:
   TDateTime ConvertLocalTimestamp(time_t Time);
   void RemoteFileTimeToDateTimeAndPrecision(const TRemoteFileTime & Source,
     TDateTime & DateTime, TModificationFmt & ModificationFmt);
-  void SetLastCode(int Code);
+  void SetLastCode(intptr_t Code);
 
   static bool Unquote(UnicodeString & Str);
   static UnicodeString ExtractStatusMessage(const UnicodeString & Status);
